Restructure the folders and add new resource for server shutdown
