+++ .gitignore
@@ -1,1 +1,3 @@
 .idea/
+src/__pycache__/
+src/resources/__pycache__/
