+++ src/credit_card.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from time import strptime
 
 
 class CreditCard:
@@ -13,8 +12,9 @@ class CreditCard:
         self.balance = self.credit_limit  # decimal positive value
 
     def pay(self, data):
-        if not data["expiration_date"] >= datetime.today():
+        if not data["expiration_date"] >= datetime.today().date():
             return {"message": f"Credit card expired on {self.expiration_date}"}
+        print(data["amount"], self.balance)
         if data["amount"] > self.balance:
             return {"message": f"Insufficient balance for the given amount "
                                f"{data['amount']}"}
